refactor: rename old footer folder
