+++ src/components/events/search/index.tsx
@@ -21,7 +21,7 @@ import { BiArrowToLeft, BiArrowToRight } from "react-icons/bi";
 import { EventCard } from "@/components/common/cards/eventCard.tsx";
 import { NavSidebar } from "@/components/events/search/nav.tsx";
 import { Sidebar } from "@/components/events/search/sidebar.tsx";
-import { FooterSection } from "@/components/footer";
+import { FooterSection } from "@/components/footerOld";
 
 export const Search = memo<{ config: any }>(({ config }) => {
   return (
